requirements.txt
Pillow==9.5.0
PyMuPDF==1.22.3
PyPDF2==3.0.1
